// src/interfaces/config.interface.ts
export interface YouSendConfig{
  token: string;
  chatId: string;
  apiKey: string;
}

export interface WapicomConfig{
  chatId: string;
  baseUrl: string;
  clientId: string;
  hash: string;
}

export interface WapicomPlConfig{
  url: string;
  hash: string;
  clientId: string;
}

export interface ServiceConfig{
  url: string;
  key: string
}